Task-07: Moved all BST classes to separate package
