+++ task-07-binary-search-tree/src/main/java/ru/otus/bst/AVLTree.java
@@ -1,4 +1,4 @@
-package ru.otus;
+package ru.otus.bst;
 
 public class AVLTree extends BinarySearchTree {
