+++ task-07-binary-search-tree/src/main/java/ru/otus/bst/BinarySearchTree.java
@@ -1,4 +1,4 @@
-package ru.otus;
+package ru.otus.bst;
 
 import static java.lang.String.format;
