+++ task-07-binary-search-tree/src/test/java/ru/otus/bst/TestHelper.java
@@ -1,4 +1,4 @@
-package ru.otus;
+package ru.otus.bst;
 
 import java.util.function.Consumer;
