+++ task-07-binary-search-tree/src/test/java/ru/otus/bst/TreapTest.java
@@ -1,4 +1,4 @@
-package ru.otus;
+package ru.otus.bst;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static ru.otus.TestHelper.checkIfKeyExists;
-import static ru.otus.TestHelper.walkTree;
+import static ru.otus.bst.TestHelper.checkIfKeyExists;
+import static ru.otus.bst.TestHelper.walkTree;
 
 class TreapTest {
